refactor(remote-fs): factor shared helpers, BlobRef, and lock-table out of the three Filesystem implementations
`in-memory.js`, `node-fs.js`, and `from-mount.js` had been
copy-pasted-then-tweaked, so each carried byte-identical copies
of several helpers. Move the duplicated pieces into
`src/shared/`:

- **`shared/helpers.js`** — pure helpers shared across all
  implementations:
  - `EMPTY_BYTES`
  - `nowNs()` / `makeAttrs()` for initial-attr records
  - `makeBytesReaderFromBytes` / `makeStringReaderFromArray` /
    `makeBytesSinkWriter` exo-stream adapters
  - `rangesOverlap(a, b)` for lock-range overlap checks
  - `assertChildName(name)` — rejects "", ".", "..", "/", NUL
  - `makeNotSupported(backing)` ENOSYS-error factory
  - `computeOpenMode(opts)` — coerces `append`/`truncate` to
    imply write (the kumavis-review fix that was previously
    triplicated)

- **`shared/blobref.js`** — `makeBlobRefExo(bytes, help?)`. One
  canonical implementation: defensively copy + harden, SHA-256,
  return the exo with `getInfo` / `fetch` / `help`. Callers can
  optionally override the `help()` body (used by `from-mount`'s
  "Mount-adapted" wording).

- **`shared/lock-table.js`** — `makeLockTable<K>()` returning
  `{ acquire(key, opts), probe(key, opts) }`. Generic over the
  key type so `in-memory.js` keys by `bigint` node-id and
  `node-fs.js` keys by absolute-path `string` with one
  implementation. The `Lock` cap that `acquire` returns is
  identical to the duplicate pair previously living in the two
  implementations.

LOC impact:

|                 | before | after | Δ     |
|-----------------|-------:|------:|------:|
| in-memory.js    |    933 |   725 |  −208 |
| node-fs.js      |    907 |   733 |  −174 |
| from-mount.js   |    677 |   607 |   −70 |
| shared/*        |      0 |   340 |  +340 |
| **net**         |  2,517 | 2,405 |  −112 |

Net 112 LOC removed, plus the more important benefit: each
shared helper now has a single canonical implementation, so
drift between the three FSes is no longer possible without
touching the shared module.

No behavioural changes. All 127 ava cases in `remote-fs` and all
18 in `9p-server` still pass; `yarn lint`, `yarn docs`, and
`yarn lint:prettier` are clean.

Some inline blocks intentionally not factored:

- `in-memory.js`'s `Directory.create()` builds a mode record
  inline because it composes `CreateFileOpts` (`{ exclusive,
  append?, truncate? }`) rather than `OpenOpts`; the shapes
  differ enough that pushing it through `computeOpenMode` would
  be a stretch.
- `node-fs.js`'s `OpenFile.write` uses a custom async sink
  iterator (each chunk awaits `fh.write`); `makeBytesSinkWriter`
  is sync-only and switching it to async would add a microtask
  hop to every in-memory write for no benefit.
- `assertChildName` is exported but not yet called by
  `in-memory.js` — in-memory historically didn't validate names,
  and adding validation would be a behavioural tightening. Left
  for a follow-up that brings the three FSes into validation
  parity.
